// app-tabman-pwa-vite/src/types/rdo-types.ts
export interface RdoMotivo {
  RDOMOTIVOCOD: number;
  DESCRICAO: string;
  SIGLA: string;
  ATIVO: 'S' | 'N';
  PRODUTIVO: 'S' | 'N';
  TOLERANCIA: number | null;
  PENALIDADE: number | null;
  WTCATEGORIA: string | null;
}
// ...
export interface DetalheFormData {
  HRINI: number;
  HRFIM: number;
  RDOMOTIVOCOD: number;
  NUOS?: number | null;
  AD_SEQUENCIA_OS?: number | null;
  CODVEICULO?: number | null;
  OBS?: string | null;
}